FFTView: added ability to change the number of amplitude bars
This allows the user to change the number of amplitude bars.

IMPORTANT NOTE: the dev could choose too many bars and the application will crash because they are too small to actually be written to the screen - or at least that's what the error looks like to me. I haven't thought of how to handle that. I'd like to have a useful fatal error to catch it and inform the dev.

Additionally, it's dawned on me that we are only showing the lowest (albiet most exciting) part of the frequency spectrum in the visualizer with 50 bars chosen. I wonder if there's a way to strategically combine bins to show more of the frequency spectrum without having too many amplitude bars..